+++ src/main/kotlin/com/estonianport/agendaza/dto/CateringEventoDto.kt
@@ -2,5 +2,5 @@ package com.estonianport.agendaza.dto
 
 import com.estonianport.agendaza.model.Extra
 
-class CateringEventoDto(val id: Long, val cateringOtro : Long, val presupuesto : Long, val descripcion : String,
+class CateringEventoDto(val id: Long, val cateringOtro : Long, val presupuesto : Double, val descripcion : String,
                         val listaExtraTipoCatering : MutableSet<ExtraDto>, val listaExtraCateringVariable : MutableSet<ExtraVariableReservaDto>) {}
