+++ src/main/kotlin/com/estonianport/agendaza/dto/EventoDto.kt
@@ -13,13 +13,13 @@ class EventoAgendaDto(var id: Long, var title: String, var start : LocalDateTime
 class EventoReservaDto(val id: Long, val nombre: String, var capacidad : Capacidad, var codigo : String,
                        val inicio : LocalDateTime, var fin : LocalDateTime, val tipoEventoId : Long,
                        val empresaId : Long, val agregados: AgregadosDto, val catering : CateringEventoDto,
-                       val cliente : Usuario, val presupuesto : Long, val encargadoId : Long, val estado : Estado) {}
+                       val cliente : Usuario, val presupuesto : Double, val encargadoId : Long, val estado : Estado) {}
 
 class EventoPagoDto(val id : Long, val nombre : String, val codigo : String,
-                      val precioTotal : Long, val listaPagos : List<PagoDto>) {}
+                      val precioTotal : Double, val listaPagos : List<PagoDto>) {}
 
 class EventoExtraDto(val id : Long, val nombre : String, val codigo : String,
-                     val presupuesto : Long, val agregados : AgregadosDto, val tipoEventoId : Long,
+                     val presupuesto : Double, val agregados : AgregadosDto, val tipoEventoId : Long,
                      val fechaEvento : LocalDateTime) {}
 
 class EventoCateringDto(val id : Long, val nombre : String, val codigo : String, val catering : CateringEventoDto,
@@ -29,7 +29,7 @@ class EventoHoraDto(val id : Long, val nombre : String, val codigo : String, val
 
 class EventoVerDto(val id : Long, val nombre : String, val codigo : String, val inicio : LocalDateTime,
                    val fin : LocalDateTime, val tipoEvento : String, val capacidad : Capacidad, val agregados : AgregadosDto,
-                   val catering : CateringEventoDto, val cliente : Usuario, val presupuesto : Long,
+                   val catering : CateringEventoDto, val cliente : Usuario, val presupuesto : Double,
                    val estado : Estado) {}
 
 class EventoBuscarFechaDto(val empresaId : Long, val desde : LocalDateTime, val hasta : LocalDateTime) {
